backend/requirements.txt
# ============================================================
# ISMS Backend - FastAPI + PostgreSQL + Celery
# ============================================================

# --- Core API Framework ---
fastapi==0.115.5
uvicorn[standard]==0.32.1

# --- Environment & Config ---
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]

# --- Email Validation ---
email-validator==2.2.0

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12

# --- Supabase ---
supabase==2.10.0

# --- HTTP Client ---
httpx==0.27.2

# --- Database (local PostgreSQL) ---
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# --- Celery (background tasks + scheduling) ---
celery==5.4.0
celery[redis]==5.4.0
flower==2.0.1

# --- Caching & Pub/Sub (Redis) ---
redis==5.2.1

# --- AI / Machine Learning ---
scikit-learn==1.6.1
numpy==2.2.3
pandas==2.2.3
anthropic>=0.40.0

# --- Utilities ---
python-dateutil==2.9.0
pytz==2024.2

# --- WebSocket (real-time) ---
websockets==13.1

# --- Firebase Cloud Messaging (push notifications) ---
firebase-admin==6.6.0